refactor(libs-backend): #0 small backend refactorings
